requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==5.0.3
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
pytest==8.0.2
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
pytest-django==4.8.0